fix: double로 변경

+++ src/main/java/com/example/planassistant/domain/Plan.java
@@ -21,8 +21,8 @@ public class Plan {
     private Long id;
     @Column(nullable = false, columnDefinition = "varChar(255) default ''")
     private String place;
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
     @Column(nullable = false, columnDefinition = "varChar(255) default ''")
     private String content;
     private LocalDateTime startTime;
@@ -33,7 +33,7 @@ public class Plan {
     @OnDelete(action = OnDeleteAction.CASCADE)
     private Member member;
     @Builder
-    public Plan(String content, String place, LocalDateTime startTime, LocalDateTime endTime, Long latitude, Long longitude, Member member) {
+    public Plan(String content, String place, LocalDateTime startTime, LocalDateTime endTime, Double latitude, Double longitude, Member member) {
         this.content = content;
         this.place = place;
         this.startTime = startTime;
